+++ singly_linked_list/src/list.h
@@ -35,7 +35,7 @@ int    list_is_empty            (List  *list);
 int    list_remove_at           (List  *list, int index);
 int    list_remove_first        (List  *list);
 int    list_remove_last         (List  *list);
-void   list_print               (List  *list, char *buffer, int len);
+void   list_print               (List  *list, char *buffer, size_t len);
 void   list_quick_sort          (List  *list);
 void   list_set                 (List  *list, int index, int value);
 int    list_size                (List  *list);
